requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pyyaml>=6.0
lxml>=5.0.0
